Fix attr interfaces up.
Rename ObjectType to AttributesType because it's about retrieving and
setting attributes, which may (in theory) be on things other than
objects some day.

Add ElementType, to access and set the element type for maps, sets, and
lists.

Add ElementsType, to access and set the element types for tuples.

Drop the SetTerraformValue method from the attr.Value interface, which
lets all our implementations go back to not being pointers again. Which
is much cleaner.